+++ core/amber/src/main/scala/edu/uci/ics/amber/engine/architecture/controller/promisehandlers/LinkWorkersHandler.scala
@@ -22,7 +22,7 @@ trait LinkWorkersHandler {
 
   registerHandler[LinkWorkers, Unit] { (msg, sender) =>
     {
-      val region = cp.workflowExecutionController.regionPlan
+      val region = cp.regionPlan
         .getRegionOfLink(msg.link)
       val resourceConfig = region.resourceConfig.get
       val linkConfig = resourceConfig.linkConfigs(msg.link)
